Upload 03-electricity-data folder
Scripts to generate the electricity data to be used in the benchmark. There are two versions, the nominal version generates incremental electricity data every 2.5 seconds to have 1440 steps in one hour. The Lower_Res instead assumes that there are 360 steps in one hour, therefore having steps with higher aplitude. In the latter case, there is a new signal every 10 seconds, which is not suited for the benchmark that works with a sampling time of 2.5 seconds. However, this dataset can be used to test preliminary approaches over more extended sets of data. In both schemes, 11 days of operations can be generated. In this repository 1 day is provided for the first approach, and 5 days for the second. The user that wants more days should modify the variable number_of_days the the respective data_import script