04-EEA-ENB-linear-matlab
This is the first version of the benchmark, for a network of linear systems and developed in Matlab. The optimization problem is solved using Gurobi, however, it can be easily converted to use Matlab built-in function quadprog, as reported in the commented part of the code. The simulation uses 24 hours of data, a different dataset needs to be used to extend it to 10 days. Additionally, there is a known error in the formulation of the integral of the tie-lines. Both data limitation, and correction if the dynamics will be performed in the next update.