Added distributed version in 04-EEA-ENB-linear-matlab
This is the version of the code implementing DMPC-ADMM for the linear version of the benchmark in Matlab. It also include genetic algorithm optimization to select partitions. In future version,Ga partitioning will be moved to a different repository and only the distributed optimization will remain in this folder. The repository needs to be checked for the consensus achievement, cause that part gave problem in the Python version, but it has been corrected and it might actually not be a problem in Matlab syntax. Also input data needs to be updated in this version.